third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/garbage-collection-and-browsing-contexts/discard_iframe_history_1-expected.txt

This is a testharness.js-based test.
[FAIL] Removing iframe from document removes it from history
  assert_equals: expected 2 but got 1
Harness: the test ran to completion.
